Add ability to parse and track CHERI context identifiers.
This patch add a new Track type for CHERI context that offer an
unstructured pid/tid pair plus a compartment identifier.
Currently the compartment identifier is an integer plus an optional
ExceptionLevel value.

Add a new metadata table for compartments, managed by the process
tracker, to complement the process and thread tables.